docker configuration

+++ generate.sh
@@ -1,32 +1,39 @@
 #!/bin/bash
 
 # Prepare environment and build package
-docker-compose down
-docker-compose up -d --build magento-test
-if [ $1 == 'true' ]
+if [ $2 == 'build' ]
+then
+    docker-compose down
+fi
+
+docker-compose up -d --build magento-$1
+if [ $1 == 'test' ]
 then
     docker-compose up -d --build selenium
 fi
 sleep 10
 
 # Install magento and sample data
-docker-compose exec magento-test install-sampledata
-docker-compose exec magento-test install-magento
+docker-compose exec -u www-data magento-$1 install-sampledata
+docker-compose exec -u www-data magento-$1 install-magento
 
 # Install modman and enable the link creation
-docker-compose exec magento-test modman init
-docker-compose exec magento-test modman link /pmt
+if [ $1 == 'test' ]
+then
+    docker-compose exec -u www-data magento-$1 modman init
+    docker-compose exec -u www-data magento-$1 modman link /pmt
+fi
 
 # Install n98-magerun to enable automatically dev:symlinks so that modman works
-docker-compose exec magento-test curl -O https://files.magerun.net/n98-magerun.phar
-docker-compose exec magento-test chmod +x n98-magerun.phar
-docker-compose exec magento-test ./n98-magerun.phar dev:symlinks 1
+docker-compose exec -u www-data magento-$1 curl -O https://files.magerun.net/n98-magerun.phar
+docker-compose exec -u www-data magento-$1 chmod +x n98-magerun.phar
+docker-compose exec -u www-data magento-$1 ./n98-magerun.phar dev:symlinks 1
 
 set -e
 # Run test
 composer install
 
-if [ $1 == 'true' ]
+if [ $1 == 'test' ]
 then
     extension/lib/DigitalOrigin/bin/phpunit --group magento-basic
     extension/lib/DigitalOrigin/bin/phpunit --group magento-configure-backoffice-iframe
@@ -37,6 +44,4 @@ then
     extension/lib/DigitalOrigin/bin/phpunit --group magento-fill-data
     extension/lib/DigitalOrigin/bin/phpunit --group magento-buy-registered
     extension/lib/DigitalOrigin/bin/phpunit --group magento-cancel-buy-registered
-else
-    extension/lib/DigitalOrigin/bin/phpunit --group magento-configure-backoffice-redirect
 fi
